+++ MediaVoyager/Clients/IGeminiRecommendationClient.cs
@@ -2,8 +2,8 @@
 {
     public interface IGeminiRecommendationClient
     {
-        public Task<string> GetMovieRecommendationAsync(List<string> favoriteMovies, List<string> watchHistory, int temperature);
+        public Task<string> GetMovieRecommendationAsync(List<string> favoriteMovies, List<string> watchHistory, double temperature);
 
-        public Task<string> GetTvShowRecommendationAsync(List<string> favouriteTvShows, List<string> watchHistory, int temperature);
+        public Task<string> GetTvShowRecommendationAsync(List<string> favouriteTvShows, List<string> watchHistory, double temperature);
     }
 }
