+++ MediaVoyager/Clients/IGroqRecommendationClient.cs
@@ -5,8 +5,8 @@ namespace MediaVoyager.Clients
 {
     public interface IGroqRecommendationClient
     {
-        public Task<string> GetMovieRecommendationAsync(List<string> favoriteMovies, List<string> watchHistory, int temperature);
+        public Task<string> GetMovieRecommendationAsync(List<string> favoriteMovies, List<string> watchHistory, double temperature);
 
-        public Task<string> GetTvShowRecommendationAsync(List<string> favouriteTvShows, List<string> watchHistory, int temperature);
+        public Task<string> GetTvShowRecommendationAsync(List<string> favouriteTvShows, List<string> watchHistory, double temperature);
     }
 }
